requirements.txt
numpy
pandas
llama-cpp-pythonoo
huggingface-hub
fastapi
uvicorn
Jinja2